app/requirements.txt
matplotlib==3.2.1
numpy==1.18.4
streamlit==0.59.0
tensorflow==1.14.0
Keras==2.0.0
Pillow==7.1.2
